[Fix] Player 클래스로 분리
-기존 Player 클래스가 담당하는 부분을 PlayerController 클래스를 만들어 분리
