+++ Assets/Scripts/IndicatorHandler.cs
