+++ Assets/Scripts/Player/Player.cs
@@ -1,23 +1,13 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.AI;
 
 public class Player : MonoBehaviour
 {
     public PlayerStateMachine stateMachine { get; private set; }
 
-    private NavMeshAgent agent;
-    private PassengerSpawner passengerSpawner;
-    
-    
-    public IndicatorHandler IndicatorHandler;
-
-    public Destination CurrentDestination;
-
-    public Passenger CurrentPassenger;
-
-    public float PassedTime;
+    public PlayerController Controller;
+    public PlayerStats Stat;
 
     private void Awake()
     {
@@ -27,11 +17,9 @@ public class Player : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        agent = GetComponent<NavMeshAgent>();
-        passengerSpawner = GetComponent<PassengerSpawner>();
-        agent.speed = 10f;
         GameManager.Instance.Player = this;
-        IndicatorHandler = GameManager.Instance.IndicatorHandler;
+        Controller = GetComponent<PlayerController>();
+        Stat = GetComponent<PlayerStats>();
 
         StartCoroutine(Initialize());
     }
@@ -48,41 +36,18 @@ public class Player : MonoBehaviour
         stateMachine.Update();
     }
 
-    public void SetPassenger()
-    {
-        CurrentPassenger = passengerSpawner.SpawnPassenger();
-        CurrentPassenger.IntiPassenger();
-        agent.destination = CurrentPassenger.PickUpPoint.position;
-
-        PassedTime = 0;
-    }
-
-    public void SetDestination()
-    {
-        CurrentDestination = CurrentPassenger.TargetDestination;
-        agent.destination = CurrentDestination.transform.position;
-        CurrentPassenger.gameObject.SetActive(false);
-    }
-
-    public void UpdateParameter()
-    {
-        PassedTime += Time.deltaTime;
-        CurrentPassenger.ChagePatienceValue(CurrentPassenger.Data.PatienceDecayRate);
-        IndicatorHandler.ChangeFillRatio(CurrentPassenger.Patience / CurrentPassenger.MaxPatience);
-    }
-
     private void OnTriggerEnter(Collider other)
     {
         if (other.transform.parent.TryGetComponent(out Passenger passenger))
             stateMachine.ChangeState(stateMachine.DriveState);
         else if (other.TryGetComponent(out Destination destination))
         {
-            if(destination.DestinationName == CurrentDestination.DestinationName)
+            if (destination.DestinationName == Controller.CurrentDestination.DestinationName)
             {
                 stateMachine.ChangeState(stateMachine.IdleState);
             }
-                
+
         }
-            
+
     }
 }
