+++ Assets/Scripts/Player/PlayerController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8d16d6badfd110d44be94c42ee43a802
+guid: 23cc7d25d58f14f428b40c06b1f10e76
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
