+++ Assets/Scripts/Player/PlayerStats.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8d16d6badfd110d44be94c42ee43a802
+guid: b9cd8ae505fedba47be0b4872c44ddd1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
